AttendanceSystem/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-decouple==3.8
numpy>=2.0.0
Pillow>=10.1.0
opencv-python>=4.8.1.78
scipy>=1.11.4
Werkzeug>=3.0.1
